define InlineBackend configurable in its own file
avoids unnecessary matplotlib imports
