+++ IPython/consoleapp.py
@@ -45,6 +45,7 @@ from IPython.kernel.zmq.kernelapp import (
     kernel_aliases,
     IPKernelApp
 )
+from IPython.kernel.zmq.pylab.config import InlineBackend
 from IPython.kernel.zmq.session import Session, default_secure
 from IPython.kernel.zmq.zmqshell import ZMQInteractiveShell
 from IPython.kernel.connect import ConnectionFileMixin
@@ -110,14 +111,7 @@ aliases.update(app_aliases)
 # IPythonConsole
 #-----------------------------------------------------------------------------
 
-classes = [IPKernelApp, ZMQInteractiveShell, KernelManager, ProfileDir, Session]
-
-try:
-    from IPython.kernel.zmq.pylab.backend_inline import InlineBackend
-except ImportError:
-    pass
-else:
-    classes.append(InlineBackend)
+classes = [IPKernelApp, ZMQInteractiveShell, KernelManager, ProfileDir, Session, InlineBackend]
 
 class IPythonConsoleApp(ConnectionFileMixin):
     name = 'ipython-console-mixin'
